+++ packages/Ecotone/src/Test/InMemoryConversionService.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace Ecotone\Messaging\Conversion;
+namespace Ecotone\Test;
 
 use Ecotone\Messaging\Config\Container\CompilableBuilder;
 use Ecotone\Messaging\Config\Container\Definition;
 use Ecotone\Messaging\Config\Container\MessagingContainerBuilder;
 use Ecotone\Messaging\Config\Container\Reference;
+use Ecotone\Messaging\Conversion\ConversionService;
+use Ecotone\Messaging\Conversion\Converter;
+use Ecotone\Messaging\Conversion\MediaType;
 use Ecotone\Messaging\Handler\Type;
 use Ecotone\Messaging\Handler\TypeDescriptor;
 use InvalidArgumentException;
