+++ packages/PdoEventSourcing/src/Prooph/EcotoneEventStoreProophWrapper.php
@@ -8,11 +8,11 @@ use DateTimeZone;
 use Ecotone\EventSourcing\EventMapper;
 use Ecotone\EventSourcing\EventStore;
 use Ecotone\Messaging\Conversion\ConversionService;
-use Ecotone\Messaging\Conversion\InMemoryConversionService;
 use Ecotone\Messaging\Conversion\MediaType;
 use Ecotone\Messaging\Handler\TypeDescriptor;
 use Ecotone\Messaging\MessageHeaders;
 use Ecotone\Modelling\Event;
+use Ecotone\Test\InMemoryConversionService;
 use Iterator;
 use Prooph\EventStore\EventStore as ProophEventStore;
 use Prooph\EventStore\Metadata\MetadataMatcher;
